Fix integer round issue

+++ java/cdk/src/main/java/sleeper/cdk/stack/CompactionStack.java
@@ -698,7 +698,7 @@ public class CompactionStack extends NestedStack {
                 //bit hacky: Reduce memory requirement for EC2 to prevent
                 //container allocation failing when we need almost entire resources
                 //of machine
-                .memoryLimitMiB(requirements.getMiddle()*0.95)
+                .memoryLimitMiB((int)(requirements.getMiddle()*0.95))
                 .gpuCount(requirements.getRight())
                 .logging(Utils.createFargateContainerLogDriver(this, instanceProperties,
                         compactionTypeName + "EC2CompactionTasks"))
